% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReviewScore.R
\name{ReviewScore}
\alias{ReviewScore}
\title{ReviewScore}
\usage{
ReviewScore(Reviews, Stars, PhraseSize = 3, nthread = 1)
}
\arguments{
\item{Reviews}{List of review text. You will likely want to remove punctuation, put everything in lowercase and other text cleaning procedures before using this function. This should usually be a column in a data frame.}

\item{Stars}{List of Star ratings associated with review. Should be on a scale of 1 to 5. This should usually be a column in a data frame.}

\item{PhraseSize}{Number stating how many words long you want the analyzed phrases to be. Defaults to 3.}

\item{nthread}{Number stating how many cores you want to run on. Defaults to 1. You may want to use more for speed. Recommended (parallel::detectCores() - 1)}
}
\value{
Data frame containing score for phrase and how often the phrase occurred in reviews.
}
\description{
This function takes reviews with star ratings and creates a score based on the frequency and positivity of reviews for phrases. Phrases that are commonly found in positive reviews will receive high scores, while phrases that frequently appear in negative reviews will have low scores. This can be useful to gain insight on important information that reviewers are happy or dissatisfied with.
}
\examples{
ReviewScore(df$ReviewText, df$StarRating, PhraseSize = 3, nthread = (parallel::detectCores() - 1))
}
